fix(ci): satisfy plugin quarantine checks
